feat(#14): Change lockfile behavior - lockfile is used only for performing backups - self-updating, config generation and validation are performed without locking - added notification of a startup error event if another nxs-backup process is already running
